+++ lib/src/l10n/croppy_localizations.dart
@@ -72,6 +72,9 @@ abstract class CroppyLocalizations {
   /// en: "Cancel"
   String get cancelLabel;
 
+  /// en: "Apply"
+  String get applyLabel;
+
   /// en: "FREEFORM"
   String get cupertinoFreeformAspectRatioLabel;