+++ lib/src/l10n/languages/croppy_zh.dart
@@ -4,6 +4,9 @@ class CroppyLocalizationsZh extends CroppyLocalizations {
   CroppyLocalizationsZh() : super('zh');
 
   @override
+  String get applyLabel => '应用';
+
+  @override
   String get cancelLabel => '取消';
 
   @override